Add default FileAttachment implementation.
It always throws not found! Also fix FileAttachments(resolve) to throw an error
synchronously, rather than asynchronously, when the file does not exist.
