+++ test/index-test.js
@@ -1,10 +1,10 @@
 import {test} from "tap";
-import {Library, FileAttachments} from "../src/index.js";
-import UID from "../src/dom/uid.js";
+import {Library} from "../src/index.js";
 
 test("new Library returns a library with the expected keys", async t => {
   t.deepEqual(Object.keys(new Library()).sort(), [
     "DOM",
+    "FileAttachment",
     "Files",
     "Generators",
     "Mutable",
@@ -18,24 +18,5 @@ test("new Library returns a library with the expected keys", async t => {
     "tex",
     "width"
   ]);
-  t.end();
-});
-
-test("FileAttachments is exported by stdlib/index", t => {
-  t.equal(typeof FileAttachments, "function");
-  t.equal(FileAttachments.name, "FileAttachments");
-  t.end();
-});
-
-test("UID", t => {
-  global.location = "https://test.com/";
-  const hi = UID("hi");
-  t.deepEqual(hi, {
-    id: "O-hi-1",
-    href: "https://test.com/#O-hi-1"
-  });
-  t.equal(hi.toString(), "url(https://test.com/#O-hi-1)");
-  const anon = UID();
-  t.equal(anon.toString(), "url(https://test.com/#O-2)");
   t.end();
 });
